Correction des warnings sous windows Ajout des conversion old-new style pour fopen, strcpy, strcat ,...

+++ src/Basic/Table.cpp
@@ -59,13 +59,13 @@ Table::~Table()
 int Table::getRowNumber() const
 {
   if (isEmpty()) return 0;
-  return _stats[0].size();
+  return static_cast<int>(_stats[0].size());
 }
 
 int Table::getColNumber() const
 {
   if (isEmpty()) return 0;
-  return _stats.size();
+  return static_cast<int>(_stats.size());
 }
 
 VectorDouble Table::getCol(int icol) const
